Move AllocationsInfo.h into the shared includes directory
This makes it possible to use AllocationsInfo by various backends.
